trading_bot/requirements.txt
upstox-api==2.3.6
yfinance==0.2.50
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.2
TA-Lib==0.4.28
vectorbt==0.25.2
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
tenacity==8.5.0
PyYAML==6.0.2
